Cosmetics/Commands/CreateToothpasteCommand.cs
using Cosmetics.Core.Contracts;
using Cosmetics.Helpers;
using Cosmetics.Models.Enums;
using System;
using System.Collections.Generic;

namespace Cosmetics.Commands
{
    public class CreateToothpasteCommand : BaseCommand
    {
        private const int ExpectedNumberOfArguments = 5;

        public CreateToothpasteCommand(IList<string> parameters, IRepository repository)
            : base(parameters, repository)
        {
        }

        public override string Execute()
        {
            ValidationHelper.ValidateArgumentsCount(this.CommandParameters, ExpectedNumberOfArguments);

            string name = CommandParameters[0];
            string brand = CommandParameters[1];
            decimal price = ParseDecimalParameter(CommandParameters[2], "price");
            GenderType gender = ParseGenderType(CommandParameters[3]);
            string ingredients = CommandParameters[4];

            if (Repository.ProductExists(name))
            {
                throw new ArgumentException("Toothpaste already exists.");
            }
            Repository.CreateToothpaste(name, brand, price, gender, ingredients);

            return $"Toothpaste with name {name} was created!";
        }

    }
}
